Same as old, but better

+++ main.py
@@ -1,11 +1,12 @@
 from datetime import datetime
 from threading import Timer
-
+import requests
+import random
 import discord
-
-from config import TOKEN
+import asyncio
 
 client = discord.Client()
+webhook = None
 
 details = {}
 
@@ -32,6 +33,8 @@ def add_info(id: int, msg: discord.Message):
     t = Timer(60, remove_first, [id])
     t.start()
 
+random_words = requests.get('https://random-word-api.herokuapp.com/all').json()
+
 
 @client.event
 async def on_ready():
@@ -40,6 +43,7 @@ async def on_ready():
 
 @client.event
 async def on_message(msg: discord.Message):
+    global webhook
     if msg.author == client.user:
         return
     user: discord.Member = msg.author
@@ -59,8 +63,27 @@ async def on_message(msg: discord.Message):
 
             add_info(user.id, msg)
             if t1 < 2 and t2 < 2:
-                await msg.delete()
+                if msg.author.display_name != 'urlocalwordspammer':
+                    await msg.delete()
+                
     else:
         add_info(user.id, msg)
-
-client.run(TOKEN)
+        
+    if '|make' in msg.content.lower():
+        channel = msg.channel
+        if webhook is None:
+            webhook = await channel.create_webhook(name='Bot Need This Thanks')
+        await msg.author.send('Ready to spam.')
+        
+    elif msg.content == '|start':
+            while True:
+                try:
+                    await webhook.send(username="urlocalwordspammer", content=f"@everyone {random.choice(random_words)}")
+                    await asyncio.sleep(1)
+                except Exception as e:
+                    await msg.author.send('Oh No. Something went wrong.')
+                    await msg.author.send(str(e))
+                    break
+        
+if __name__ == '__main__':
+    client.run('ODI3Mzk3OTc2MDc0NDg1Nzgw.YGacaQ.h5MapmWuPuETUnVXQsCgHE3BtGo')
